Add data in database; Update home page; Update movie service

+++ src/main/java/mk/ukim/finki/wp/cineverse/service/MovieService.java
@@ -14,9 +14,9 @@ public interface MovieService {
     Movie removeById(Long id);
     Movie save(String title, String genre, String description,
                Image poster, String trailerUrl, String showtime, String duration,
-               Long ticketPrice, String director, String writer, List<Actor> actors);
+               Float ticketPrice, String director, String writer, List<Actor> actors);
     Movie update(Long movieId, String title, String genre, String description,
                  Image poster, String trailerUrl, String showtime, String duration,
-                 Long ticketPrice, String director, String writer, List<Actor> actors);
+                 Float ticketPrice, String director, String writer, List<Actor> actors);
     List<Movie> getLatestMovies();
 }
